refactor: participationrequest 패키지명 groupapplication으로 변경
