+++ main/src/main/java/com/amorgakco/backend/group/service/GroupService.java
@@ -10,7 +10,7 @@ import com.amorgakco.backend.group.dto.GroupRegisterRequest;
 import com.amorgakco.backend.group.repository.GroupRepository;
 import com.amorgakco.backend.group.service.mapper.GroupMapper;
 import com.amorgakco.backend.member.domain.Member;
-import com.amorgakco.backend.participationrequest.repository.GroupApplicationRepository;
+import com.amorgakco.backend.groupapplication.repository.GroupApplicationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
