+++ main/src/main/java/com/amorgakco/backend/groupapplication/service/mapper/GroupApplicationMapper.java
@@ -1,7 +1,7 @@
-package com.amorgakco.backend.participationrequest.service.mapper;
+package com.amorgakco.backend.groupapplication.service.mapper;
 
 import com.amorgakco.backend.group.domain.Group;
-import com.amorgakco.backend.participationrequest.domain.GroupApplication;
+import com.amorgakco.backend.groupapplication.domain.GroupApplication;
 import com.amorgakco.backend.member.domain.Member;
 import org.springframework.stereotype.Component;
