+++ main/src/test/java/com/amorgakco/backend/fixture/groupparticipation/TestGroupParticipationFactory.java
@@ -3,7 +3,7 @@ package com.amorgakco.backend.fixture.groupparticipation;
 import com.amorgakco.backend.fixture.group.TestGroupFactory;
 import com.amorgakco.backend.fixture.member.TestMemberFactory;
 import com.amorgakco.backend.group.domain.Group;
-import com.amorgakco.backend.participationrequest.domain.GroupApplication;
+import com.amorgakco.backend.groupapplication.domain.GroupApplication;
 import com.amorgakco.backend.member.domain.Member;
 
 public class TestGroupParticipationFactory {
